+++ src/plugins/relay/relay-http.h
@@ -40,6 +40,7 @@ enum t_relay_client_http_status
 #define RELAY_HTTP_401_UNAUTHORIZED           401, "Unauthorized"
 #define RELAY_HTTP_403_FORBIDDEN              403, "Forbidden"
 #define RELAY_HTTP_404_NOT_FOUND              404, "Not Found"
+#define RELAY_HTTP_405_METHOD_NOT_ALLOWED     405, "Method Not Allowed"
 #define RELAY_HTTP_500_INTERNAL_SERVER_ERROR  500, "Internal Server Error"
 #define RELAY_HTTP_503_SERVICE_UNAVAILABLE    503, "Service Unavailable"
 
@@ -53,6 +54,7 @@ enum t_relay_client_http_status
 #define RELAY_HTTP_ERROR_INVALID_ITERATIONS   "Invalid number of iterations"
 #define RELAY_HTTP_ERROR_BAD_REQUEST          "Bad request"
 #define RELAY_HTTP_ERROR_NOT_FOUND            "Resource not found"
+#define RELAY_HTTP_ERROR_METHOD_NOT_ALLOWED   "Method Not Allowed"
 #define RELAY_HTTP_ERROR_OUT_OF_MEMORY        "Out of memory"
 
 struct t_relay_http_request